airflow/requirements.txt
kafka-python==2.1.2
praw==7.8.1
apache-airflow-providers-docker==4.0.0
pymongo==4.5.0 